+++ pyvo/samp/constants.py
