+++ pyvo/samp/data/astropy_icon.png
