+++ pyvo/samp/errors.py
