+++ setup.cfg
@@ -89,7 +89,7 @@ pyvo.registry.tests = data/*.xml, data/*.desise
 pyvo.mivot.writer = *.xsd
 pyvo.mivot.tests = data/*.xml, data/input/*.xml, data/output/*.xml, data/reference/*json, data/reference/*xml
 pyvo.dal.tests = data/*.xml, data/*/*
-pyvo.astropy_samp = data/astropy_icon.png, data/*xml
+pyvo.samp = data/astropy_icon.png, data/*xml
 
 [coverage:run]
 source = pyvo
